--- client/src/components/CRUD/addBlog/addBlogForm.js
@@ -1,14 +1,0 @@
-import AddPost from "./addPost";
-import AddAuthor from "./addAuthor";
-
-const AddBlogForm = () => {
-
-  return(
-<div>
-  This is from the AddBlogForm component.
-<AddPost />
-<AddAuthor />
-</div>
-  )
-}
-export default AddBlogForm;
