+++ client/src/components/CRUD/addBlog/addPost.js
@@ -1,63 +1,72 @@
 import { useState } from 'react';
 
 
-const AddPost = ({ addPostCallback }) => {
+const AddPost = (props) => {
+  //POSTS
   const [postForm, setPostForm] = useState({
     title: "",
     date: "",
     post: "",
+    id: ""
   });
 
-
-  //POSTS - create functions that handle the event of the user typing into the form
-  const handleTitle = (event) => {
-    const postTitle = event.target.value;
+  const handleTitle = (e) => {
+    const postTitle = e.target.value;
     setPostForm((post) => ({ ...post, title: postTitle }));
   };
 
-  const handleDate = (event) => {
-    const postDate = event.target.value;
+  const handleDate = (e) => {
+    const postDate = e.target.value;
     setPostForm((post) => ({ ...post, date: postDate }));
   };
-  const handlePost = (event) => {
-    const postPost = event.target.value;
+
+  const handlePost = (e) => {
+    const postPost = e.target.value;
     setPostForm((post) => ({ ...post, post: postPost }));
   };
 
-
-//POST Request - Posts
-  const postPost = (newPost) => {
-    return fetch("http://localhost:8080/posts", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify(newPost),
-    })
-      .then((response) => {
-        return response.json();
-      })
-      .then((data) => {
-        console.log("From the post ", data);
-        addPostCallback.addPost(data);
-      });
+  const handleId = (e) => {
+    const postId = e.target.value;
+    setPostForm((post) => ({ ...post, id: postId }));
   };
 
-//addPost will be the props
-  const handleAddPost = (e) => {
+  //AUTHORS
+  const [authorForm, setAuthorForm] = useState({
+    name: '',
+    post_id: ''
+  });
+
+  const handleName = (e) => {
+    const authorName = e.target.value;
+    setAuthorForm((author) => ({ ...author, name: authorName }));
+  };
+  const handlePostId = (e) => {
+    const authorPostId = e.target.value;
+    setAuthorForm((author) => ({ ...author, post_id: authorPostId }));
+  };
+
+  //EVENT HANDLER
+  const handleAddBlogEntry = (e) => {
     e.preventDefault();
-    postPost(postForm);
+
+    console.log(postForm);
+    console.log(authorForm);
+    props.addNewSubmission(postForm, authorForm);
     setPostForm({
       title: "",
       date: "",
-      post: ""
+      post: "",
+      id:""
     })
   };
-  
+
 
   return (
-    <form onSubmit={handleAddPost} className="formborder" >
+    <form onSubmit={handleAddBlogEntry} className="formborder" >
       <fieldset>
+        {/* ******************** POST - postForm ******************* */}
 
-      <label>Title: </label>
+        <label className='addlabel'>Title: </label>
         <input
           type="text"
           id="add-title"
@@ -67,17 +76,7 @@ const AddPost = ({ addPostCallback }) => {
         />
         <br />
 
-        <label>Post: </label>
-        <input
-          type="text"
-          id="add-post"
-          placeholder="Blog text goes here."
-          value={postForm.post}
-          onChange={handlePost}
-        />
-        <br />
-
-        <label>Date: </label>
+        <label className='addlabel'>Date: </label>
         <input
           type="date"
           id="add-date"
@@ -86,8 +85,51 @@ const AddPost = ({ addPostCallback }) => {
         />
         <br />
 
+        <label className='addlabel'>Post: </label>
+        <textarea
+          className='postentryinput'
+          type="text"
+          id="add-post"
+          placeholder="Blog text goes here."
+          value={postForm.post}
+          onChange={handlePost}
+        />
+        <br />
+
+        <label className='addlabel'>Post ID: </label>
+        <input
+          type="text"
+          id="add-postId"
+          placeholder="113"
+          value={postForm.id}
+          onChange={handleId}
+        />
+        <br />
+        {/* ******************** AUTHOR - authorForm ******************* */}
+        <label className='addlabel'>Author: </label>
+        <input
+          type="text"
+          id="add-author"
+          placeholder="Jane Smith"
+          required
+          value={authorForm.name}
+          onChange={handleName}
+        />
+        <br />
+
+        <label className='addlabel'>Author ID: </label>
+        <input
+          type="text"
+          id="add-postID"
+          placeholder="4"
+          required
+          value={authorForm.post_id}
+          onChange={handlePostId}
+        />
+        <br />
+
       </fieldset>
-      <button type="submit">Add</button>
+      <button className="addpostbutton" type="submit"> Add Post </button>
     </form>
   );
 };
